Move merchant center link examples to AM directory
Matches location of other libs
